all requests are now in requests file for meals, event handlers are in event handler file
