+++ lib/index.js
@@ -18,12 +18,13 @@ const updateDiary = require('./meals/updateDiary')
 const mealFilter = require('./meals/filter')
 const addMeal = require('./meals/addMeal')
 const sortFoods = require('./meals/sortFoods')
+const mealEventHandler = require('./meals/eventHandler')
 
 $(document).ready(() => {
   mealRequests.getMeals();
   $('.new-food-btn').on('click', updateDiary.redirectNewFood)
   $('#diaryFilter').on('keyup', mealFilter.filterDiary);
   $('.add-meal-btn').on('click', addMeal.getCheckBoxIds);
-  $('.all-meals').on('click', '.delete-meal-btn', mealRequests.deleteMeal);
+  $('.all-meals').on('click', '.delete-meal-btn', mealEventHandler.deleteMeal);
   $('#calorie-sort').on('click', sortFoods.sortFoods)
 });
