+++ lib/meals/addMeal.js
@@ -1,4 +1,5 @@
 const requests = require('./requests')
+const eventHandler = require('./eventHandler')
 
 export function getCheckBoxIds() {
   event.preventDefault();
@@ -8,7 +9,7 @@ export function getCheckBoxIds() {
   checkedFoods.forEach(function(id) {
     for (var i = 0; i < diaries.length; i++) { 
       if (diaries[i].dataset.foodId === id) { 
-        requests.addFoodToMeal(diaries[i], meal) 
+        eventHandler.addFoodToMeal(diaries[i], meal) 
       } 
     }
   })
