+++ lib/meals/requests.js
@@ -3,37 +3,25 @@ const addMeal = require('./addMeal')
 const totalsTable = require('./totalsTable')
 const updateDiary = require('./updateDiary')
 const APIhost = `https://evening-peak-25213.herokuapp.com`
+const eventHandler = require('./eventHandler')
 
 export const getMeals = () => {
   return fetch(APIhost + `/api/v1/meals`)
     .then(response => handleResponse.handleResponse
       (response))
-    .then(meals => getAllMeals(meals))
+    .then(meals => eventHandler.getAllMeals(meals))
     .catch(error => console.log({ error }));
 }
 
 export const getFoodsDiary = () => {
   return fetch(APIhost + `/api/v1/foods`)
     .then(response => handleResponse.handleResponse(response))
-    .then(foods => sortById(foods))
-    .then(sortedFoods => getAllFoodsDiary(sortedFoods))
+    .then(foods => eventHandler.sortById(foods))
+    .then(sortedFoods => eventHandler.getAllFoodsDiary(sortedFoods))
     .catch(error => console.log({ error }));
 }
 
-const sortById = (unsorted) => {
-  let unsortedArray = [...unsorted]
-  const sorted = unsortedArray.sort((firstArticle, secondArticle) => {
-    return firstArticle.id - secondArticle.id
-  })
-  return sorted
-}
-
-export const addFoodToMeal = (food, meal) => {
-  const foodObject = { name: food.getElementsByTagName('p')[0].innerText, 
-                     calories: parseInt(food.getElementsByTagName('p')[1].innerText), 
-                     id: food.dataset.foodId }
-  const data = { name: foodObject.name, calories: foodObject.calories }
-  const mealId = addMeal.getMealId(meal);
+export const addFoodToMealRequest = (foodObject, data, mealId, food, meal) => {
   return fetch(APIhost + `/api/v1/meals/` + mealId + `/foods/` + foodObject.id, {
     method: 'post',
     headers:
@@ -47,11 +35,7 @@ export const addFoodToMeal = (food, meal) => {
     .catch(error => console.log({ error }))
 };
 
-export function deleteMeal(event) {
-  const food = event.currentTarget.parentElement.parentElement;
-  const mealName = food.className.toLowerCase()
-  const mealId = addMeal.getMealId(mealName)
-  const foodId = food.getAttribute(`data-${mealName}-food-id`)
+export function deleteMealRequest(food, mealName, mealId, foodId) {
   return fetch(APIhost + `/api/v1/meals/` + mealId + `/foods/` + foodId, {
     method: 'delete',
     headers:
@@ -60,24 +44,4 @@ export function deleteMeal(event) {
     .then(food.remove())
     .then(updateDiary.updateTotalsPostDelete(food, mealName))
     .catch(error => console.log({ error }))
-}
-
-export const getAllFoodsDiary = (foods) => {
-  return foods.forEach((food) => {
-    addMeal.prependFoodDiary(food)
-  })
-}
-
-const getAllMeals = (meals) => {
-  getFoodsDiary();
-  getTotalsTable(meals);
-  return meals.forEach((meal) => {
-    addMeal.prependMeal(meal);
-  })
-}
-
-const getTotalsTable = (meals) => {
-  totalsTable.appendToTotalCaloriesConsumed(meals);
-  totalsTable.appendToTotalGoalCalories(meals);
-  totalsTable.appendToTotalCaloriesRemaining(meals);
 }
